fix: Skeleton Component & relay test
