+++ src/pages/index.tsx
@@ -1,16 +1,9 @@
 import { keys, map, pipe, toArray } from '@fxts/core';
-import { usePreloadedQuery } from 'react-relay';
-import { RelayProps, withRelay } from 'relay-nextjs';
 
 import { AsyncBoundary } from 'src/components/AsyncBoundary';
 import { Typography, Toggle, Popup, BottomTab, Skeleton } from 'src/ui';
 import Icon from 'src/ui/Icon/Icon';
 import { SVG_ICON_MAP } from 'src/ui/Icon/iconMap';
-import { getClientEnvironment } from 'src/utils/relay/client_environment';
-import { createServerEnvironment } from 'src/utils/relay/server_environment';
-
-import { query } from '../components/RelayTest';
-import { RelayTestQuery } from '../queries/__generated__/RelayTestQuery.graphql';
 
 const LoadingComponent = () => <h1>...loading</h1>;
 const ErrorFallback = () => (
@@ -29,22 +22,12 @@ const ErrorFallback = () => (
   </div>
 );
 
-const Test = ({ preloadedQuery }: Omit<RelayProps<{}, RelayTestQuery>, 'CSN'>) => {
-  const data = usePreloadedQuery(query, preloadedQuery);
-  return <div>{data.allFilms?.films?.[0]?.title}</div>;
-};
-
-const Home = ({ preloadedQuery }: RelayProps<{}, RelayTestQuery>) => (
+const Home = () => (
   <AsyncBoundary errorFallback={ErrorFallback} loadingFallback={LoadingComponent}>
-    <Test preloadedQuery={preloadedQuery} />
     <Skeleton height="sm" width="full" />
     <Skeleton borderRadius="md" height="lg" width="lg" />
     <BottomTab />
   </AsyncBoundary>
 );
 
-export default withRelay(Home, query, {
-  createClientEnvironment: () => getClientEnvironment()!,
-  createServerEnvironment: async () => createServerEnvironment(),
-  variablesFromContext: () => ({ first: 1 }),
-});
+export default Home;
